wayfair_clicker/pyproject.toml
[project]
name = "wayfair"
version = "0.1.0"
description = ""
authors = [
    {name = "v.kondratov"}
]

readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "telebot (>=0.0.5,<0.0.6)",
    "python-decouple (>=3.8,<4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
